[:hammer: refactor #3] modifica el getter del admin
