+++ _funcs/getSO.py
@@ -1,12 +1,6 @@
-import re, subprocess, ctypes, os, sys, platform
 from _utils.bannerEmmiter import bannerEmmiter
-def is_admin():
-    is_admin = False
-    try:
-        is_admin = os.getuid() == 0
-    except AttributeError:
-        is_admin = ctypes.windll.shell32.IsUserAnAdmin() != 0
-    return is_admin
+import sys, re, subprocess
+from _utils.adminValidator import is_admin
 
 def getTtl(host):
     # Se obtiene el TTL del host
