+++ _utils/adminValidator.py
@@ -1,0 +1,8 @@
+import ctypes, os
+def is_admin():
+    is_admin = False
+    try:
+        is_admin = os.getuid() == 0
+    except AttributeError:
+        is_admin = ctypes.windll.shell32.IsUserAnAdmin() != 0
+    return is_admin
