+++ main.py
@@ -1,10 +1,11 @@
 import argparse, signal
-from _utils  import handlerCtrsC
+from _utils import handlerCtrsC
 
 # Función para capturar la señal SIGINT (Ctrl+C)
 signal.signal(signal.SIGINT, handlerCtrsC.def_handler)
 
 def main():
+
   parser = argparse.ArgumentParser(description="""
     ______   ______   _   __  ______   ______   
    |_    _| |  __  | | | / / |  ____| |  __  |  
@@ -29,7 +30,7 @@ def main():
 
   if args.scanPorts:
     if args.scanPorts:
-      import scanPorts
+      from _funcs import scanPorts
       scanPorts.scanPorts()
 
   if args.crackMD5:
